Remove the redundant interface.
This class declares that it implements an interface Map which is also implemented by the superclass HashMap.
This is redundant because once a superclass implements an interface, all subclasses by default also implement this interface. It may point out that the inheritance hierarchy has changed since this class was created, and consideration should be given to the ownership of the interface's implementation.
http://findbugs.sourceforge.net/bugDescriptions.html#RI_REDUNDANT_INTERFACES